solved: prefix for subscribing topics is added to additional subscriptions #35
A new switch has been installed which specifies whether the prefix is ​​used for additional subscriptions.
If the switch is activated, all statuses of the MQTT Client namespaces are processed without a prefix.
If the switch is set, the system behaves as before.

Further corrections:

If a status was added by an additional subscription, the value was missing. This is now initially added as text.

When changing the topic of a subscripted status, the old topic was not removed from the subscriptions.

When deleting a subscripted status, the old topic was not removed from the subscriptions.

If the prefix was changed in the config, the topic was also subscribed with the old prefix.

If an additional subscription receives a value for a non-subscripted value, the value of the status has been changed.

Unsubscribe did not work with a prefix.